refactor slider
