+++ src/components/dialogs/card-dialog/card-dialog.stories.tsx
@@ -1,9 +1,10 @@
 import { useState } from 'react'
 
 import defaultImage from '@/assets/images/cover.png'
-import { CardDialog } from '@/components/dialogs'
 import { Button } from '@/components/ui'
 import { Meta, StoryObj } from '@storybook/react'
+
+import { CardDialog } from './card-dialog'
 
 const meta = {
   component: CardDialog,
