+++ src/components/ui/index.ts
@@ -1,16 +1,18 @@
-export * from './button/button'
-export * from './checkbox/checkbox'
-export * from './input/input'
-export * from './tab-switcher/tab-switcher'
-export * from './table/table-primitives/table'
-export * from './typography/typography'
-export * from './card/card'
-export * from './radio-group/radio-group'
-export * from './select'
-export * from './modal/modal'
 export * from './avatar/avatar'
-export * from './grade-icons/grade-icons'
-export * from './pagination/pagination'
-export * from './table-sort-button/table-sort-button'
-export * from './dialog/dialog'
 export * from './back-link/back-link'
+export * from './button/button'
+export * from './card/card'
+export * from './checkbox/checkbox'
+export * from './dialog/dialog'
+export * from './drop-down/drop-down'
+export * from './grade-icons/grade-icons'
+export * from './input/input'
+export * from './modal/modal'
+export * from './pagination/pagination'
+export * from './radio-group/radio-group'
+export * from './select/select'
+export * from './slider/slider'
+export * from './table-sort-button/table-sort-button'
+export * from './table/table-primitives/table'
+export * from './tab-switcher/tab-switcher'
+export * from './typography/typography'
