--- src/components/ui/slider/index.ts
@@ -1,1 +1,0 @@
-export * from './slider'
